% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anytimetest.R
\docType{package}
\name{anytimetest-package}
\alias{anytimetest}
\alias{anytimetest-package}
\title{anytimetest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com}

}
\keyword{internal}
