% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utctime2.R
\name{utctime2}
\alias{utctime2}
\title{Test anytime}
\usage{
utctime2(x)
}
\arguments{
\item{x}{A vector of type character, integer or numeric with date(time) expressions to be parsed and converted.}
}
\description{
Test anytime
}
